requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Configuration
pydantic-settings==2.5.2

# Task Queue
celery[redis]==5.4.0
redis==5.1.1

# Encryption
cryptography==43.0.1

# PDF Parsing
PyMuPDF==1.24.10
pdfplumber==0.11.4

# LLM Providers
openai==1.51.0
anthropic==0.34.2
httpx==0.27.2

# Browser Automation
playwright==1.47.0

# LaTeX Templating
Jinja2==3.1.4

# Validation
email-validator==2.2.0

# Utilities
python-dateutil==2.9.0
aiofiles==24.1.0
